Update repo to upstream (#3)
* Update faq.md

* Fixed spelling errors

closes #13

* Updated to include @ksmai's app for practice

Closes #11

* Typo built -> build

* Add third chapter of git guide to faq

* Update FAQ to include helping others.

Add some advice about the positive effects on your learning when you take the time to explain topics to someone else.

* Replace MongoDB course

Confirmed with MongoDB University that the edX course is retired.

Replaced it in the curriculum with two options, learners may
select one, the other, or both. One is more full stack oriented,
the other dives into backend a bit more. Both are worthwhile, either
is sufficient to prep for the FCC backend projects.

Updated CHANGELOG to reflect curriculum change.

* job-ready.md: typo you're -> your

* you're -> your

* Add Data Science Guide

* Add Data Science Guide

* Modify instructions for MongoDB Performance course to reflect that it's in Tier 2

* Modified Suggestions

Added suggestion related to ensuring both Python and R are each
used on a variety of projects to ensure competancy in both by
the end of the guide.

* Add Python Guide

* Rearrange Python guide to be at top level

* Fix link to Python Guide

* Refine Python Guide Projects

* Refine Python Guide Projects

* Add sql to Python Guide

* Fix links in What Now section.

* Added Django as an option to Flask

* Add Django as an option to Flask

* Add 2.0 version of JavaScript Job Ready Guide

* Add 2.0 version of JavaScript Job Ready Guide

* Slight styling change on main README

* fix wayward apostrophes

* Fix typo

addtional -> additional

* Update job-ready-python-edition.md

Typo - ideal was typed twice.

* fix spelling

i saw it so i PR-ed it.

* Move MongoDB to the beginning of Tier 2 / clarify.

The MongoDB course is offered on a schedule and has assignments that unfold over the course of 7 weeks. Moved signing up for it to the beginning of Tier 2, with instructions to begin it during the next available session. Originally, this was an either/or course with Programming for the Web, but I've moved the latter into Tier 1 and am considering them a two-course sequence now as one focuses more on frontend, the other on backend. 

Closes #25

* Update job-ready-python-edition.md

I was clicking on these titles in Contents and the links weren't working. So, I removed the hyphen in the end of the links.

* Update job-ready-javascript-edition-2.0.md

* Rearrange assets

* Fix typo

Typo in "employeers"

* Update Chingu Cohort Link

* Added Software Construction: Object-Oriented Design

Added the next course in the How to Code series to the CS-WD and WD-CS guides. My intent is to include all courses in the series (they are fantastic) but this one just became available this week and the final one isn't available until 2018, so I'll add that one at that time.

* Modify project activities to include additional options.

* Fix spelling.

* corrected spelling of "glitch"

* Change data visualization course

Replaced the now paid data visualization course from egghead with a free course on udacity

* Replace D3 course with free alternative

* Replace D3 course with free alternative.

* Swapped out paid D3 course for free alternative.

* Update spelling of accomplished

* Fix typo

* Add Practice/Speedrun  project/problem lists.

* Fix headings on new Speedrun / Practice lists.

* Remove extraneous characters from intro list.

* Re-add missing items to frontend list.

* Add index to practice speedrun directory.

* fix typo in fullstack-dev.md

* Alternative spelling for a few words

relevent -> relevant
Algorthms -> Algorithms
intermitantly -> intermittently

* Update faq.md

fix typo

* Add next course in the How to Code Series, plus link to the Speedstudy repo.

* Update cs-wd.md

Getting Sassy with SaSS to Getting Sassy with CSS

* Add Angular Guide.

* Add Vue topic guide.

* Add React topic guide. Refactor Vue and Angular guides for consistency between the three.

* Modify testing instructions for Tier 2 for consistency with where Jest/Enzyme appear in the sequence.

* Update project instructions on React Guide.

* Change Tutorial to Official Docs where appropriate in Topic Guides.

* Tweak instructions for each Tier in Topic Guides.

* Update to include mobile app in Bonus section.

* Include Flexbox/Grid in Tier 1 of frontend framework topic guides.

* Add RxJS to topic guides for frontend frameworks.

* Slight tweak to Angular Tier 4 to make it more consistent.

* Fix CodeWars links in Angular/React/Vue topic guides. Closes #50.

* Add Flow to Bonus - advanced section.

* Rectify a typo

* Fix spelling.

* Fix spelling.

* Corrected spelling

Corrected spelling of plethora, include & fulfill

* Update C++, Add JavaScript.

* Update interview prep

* Update JavaScript Prep

* Minor typo changes in React guide

* Add Software Development Capstone Project to CSWD and WDCS guides.

* Update Calculus Courses.
OSUs Coursera Calculus courses are being shut down, switched to MITOCW.

* Fix layout on the ubcx new courses added.

* Add CS/Java Focus Guide

* Replace defunct Natural Language Processing course.

* Replace defunct Natural Language Processing course.

* Update Changelog.

* Fix typo 'inculde' to 'include'

* Update FAQ links to properly work

* Type job ready

Changed 'pleathora' to plethora.

* Fix double spacing typos

* Fix regex deep dive subheading's title typo